Extract enum level value adapter
