+++ library/src/main/java/com/github/mproberts/rxdatabinding/bindings/EnumLevelValueAdapter.java
@@ -1,0 +1,20 @@
+package com.github.mproberts.rxdatabinding.bindings;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class EnumLevelValueAdapter<T> implements LevelValueAdapter {
+
+    Map<T, Integer> _enumValues = new HashMap<>();
+
+    @Override
+    public int convertValue(Object value) {
+        return _enumValues.get((T) value);
+    }
+
+    public EnumLevelValueAdapter<T> map(T enumValue, int level) {
+        _enumValues.put(enumValue, level);
+
+        return this;
+    }
+}
